MeanMaximization.java
Mean Maximization Problem Code: MEANMAXSolvedSubmit
JJ loves playing with averages. He has an array A of length N. He wants to divide the array A into two non-empty subsets P and Q such that the value of mean(P)+mean(Q) is as large as possible. (Note that each Ai must belong to either subset P or subset Q).

Help him find this maximum value of mean(P)+mean(Q).

As a reminder, the mean of a subset X of size M is defined as: mean(X)=∑Mi=1XiM.

For example, mean([3,1,4,5])=3+1+4+54=3.25.